requirements.txt
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
pillow==9.5.0
tqdm==4.65.0
